[BAU] refactor schema validation
