+++ tests/validation_tests/schema_validation_tests/test_schema.py
@@ -3,7 +3,7 @@ from enum import StrEnum
 
 import pytest
 
-from core.validation.schema import SchemaError, parse_schema
+from core.validation.schema_validation.schema import SchemaError, parse_schema
 
 ####################################
 # Test parse_schema
